Update and rename Different test to Different test in C
Description:

This code demonstrates the use of pointers and arrays in C. It defines an array of integers called "names" and a pointer "pn" that points to the first element of the array. The code then prints the memory addresses and sizes of both the array elements and the pointer elements using a loop. Finally, the code prints the memory address of the entire array.