Update and rename project1 to RandomNumberGuesser
Update Description:

- Improved formatting and indentation of the code to make it easier to read
- Added function prototypes and docstrings to describe the purpose and parameters of the `bigger` and `lower` functions
- Changed the `answer` and `answerYN` arrays to be of type `char` instead of `int` to match the expected input from the user
- Added a check to ensure that the user enters a valid number within the specified range

Description:

This code is a simple number guessing game that asks the user to enter a number between 0 and 100. The program then generates a random number and asks the user if it is the same as the number they entered. If the numbers match, the game ends. If the numbers do not match, the program asks the user if the number they entered is bigger or smaller than the generated number, and adjusts the range accordingly. The game continues until the program guesses the correct number.